break out topiccomment from main dao into separate dao
