+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/SdelDao.java
@@ -34,9 +34,6 @@ public interface SdelDao {
 	Image addImage(Image image);
 	
 
-	
-	TopicComment addTopicComment(TopicComment topicComment);
-	
-	TopicComment updateTopicComment(TopicComment topicComment);
+
 	
 }
