+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/TopicCommentDAO.java
@@ -1,0 +1,11 @@
+package com.skilldistillery.sdelp.data;
+
+import com.skilldistillery.sdelp.entities.TopicComment;
+
+public interface TopicCommentDAO {
+	
+	TopicComment addTopicComment(TopicComment topicComment);
+	
+	TopicComment updateTopicComment(TopicComment topicComment);
+
+}
